requirements.txt
botorch>=0.8.0
gpytorch>=1.11.0
torch>=2.0.0
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.5.0
google-genai
pyyaml
ConfigSpace
